fix(tracemetrics): Remount filter bar when metric changes
