+++ static/app/views/explore/metrics/metricToolbar/aggregateDropdown.tsx
