+++ static/app/views/explore/metrics/metricToolbar/deleteMetricButton.tsx
