+++ static/app/views/explore/metrics/metricToolbar/groupBySelector.tsx
