+++ static/app/views/explore/metrics/metricToolbar/metricSelector.tsx
